\name{print.modTempEff}
\alias{print.modTempEff}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{The print method for a 'modTempEff' object}
\description{
Prints a Constrained Segmented Distributed Lag Model object  
}
\usage{
\method{print}{modTempEff}(x, digits = max(3, getOption("digits") - 3), ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ object of class \code{modTempEff} }
  \item{digits}{ number of digits to be printed }
  \item{\dots}{ additional arguments.. }
}
\details{
  Prints the minimal features of a \code{modTempEff} object.
}
%\references{ ~put references to the literature/web site here ~ }
\author{ Vito Muggeo}
\seealso{ \code{\link{summary.modTempEff}} }
\keyword{ regression}
